refactor(server): reorganise break the monolith dtos
