+++ packages/amplication-server/src/core/ai/dto/index.ts
@@ -1,0 +1,3 @@
+export * from "./BtmRecommendation";
+export * from "./BtmResourceRecommendation";
+export * from "./BtmEntityRecommendation";
